+++ lib/src/service/media/uploaded_media_data.freezed.dart
@@ -32,10 +32,10 @@ mixin _$UploadedMediaData {
   /// The date and time this media will expire.
   DateTime get expiresAt => throw _privateConstructorUsedError;
 
-  /// Indicates the language of this media.
+  /// The locale of this media.
   ///
-  /// This field is only set if a caption (.srt) is uploaded.
-  Language? get language => throw _privateConstructorUsedError;
+  ///
+  Locale? get locale => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
@@ -55,7 +55,9 @@ abstract class $UploadedMediaDataCopyWith<$Res> {
       @JsonKey(name: 'media_id_string')
           String id,
       DateTime expiresAt,
-      Language? language});
+      Locale? locale});
+
+  $LocaleCopyWith<$Res>? get locale;
 }
 
 /// @nodoc
@@ -72,7 +74,7 @@ class _$UploadedMediaDataCopyWithImpl<$Res>
     Object? mediaId = freezed,
     Object? id = freezed,
     Object? expiresAt = freezed,
-    Object? language = freezed,
+    Object? locale = freezed,
   }) {
     return _then(_value.copyWith(
       mediaId: mediaId == freezed
@@ -87,11 +89,22 @@ class _$UploadedMediaDataCopyWithImpl<$Res>
           ? _value.expiresAt
           : expiresAt // ignore: cast_nullable_to_non_nullable
               as DateTime,
-      language: language == freezed
-          ? _value.language
-          : language // ignore: cast_nullable_to_non_nullable
-              as Language?,
+      locale: locale == freezed
+          ? _value.locale
+          : locale // ignore: cast_nullable_to_non_nullable
+              as Locale?,
     ));
+  }
+
+  @override
+  $LocaleCopyWith<$Res>? get locale {
+    if (_value.locale == null) {
+      return null;
+    }
+
+    return $LocaleCopyWith<$Res>(_value.locale!, (value) {
+      return _then(_value.copyWith(locale: value));
+    });
   }
 }
 
@@ -109,7 +122,10 @@ abstract class _$$_UploadedMediaDataCopyWith<$Res>
       @JsonKey(name: 'media_id_string')
           String id,
       DateTime expiresAt,
-      Language? language});
+      Locale? locale});
+
+  @override
+  $LocaleCopyWith<$Res>? get locale;
 }
 
 /// @nodoc
@@ -128,7 +144,7 @@ class __$$_UploadedMediaDataCopyWithImpl<$Res>
     Object? mediaId = freezed,
     Object? id = freezed,
     Object? expiresAt = freezed,
-    Object? language = freezed,
+    Object? locale = freezed,
   }) {
     return _then(_$_UploadedMediaData(
       mediaId: mediaId == freezed
@@ -143,10 +159,10 @@ class __$$_UploadedMediaDataCopyWithImpl<$Res>
           ? _value.expiresAt
           : expiresAt // ignore: cast_nullable_to_non_nullable
               as DateTime,
-      language: language == freezed
-          ? _value.language
-          : language // ignore: cast_nullable_to_non_nullable
-              as Language?,
+      locale: locale == freezed
+          ? _value.locale
+          : locale // ignore: cast_nullable_to_non_nullable
+              as Locale?,
     ));
   }
 }
@@ -161,7 +177,7 @@ class _$_UploadedMediaData implements _UploadedMediaData {
       @JsonKey(name: 'media_id_string')
           required this.id,
       required this.expiresAt,
-      this.language});
+      this.locale});
 
   factory _$_UploadedMediaData.fromJson(Map<String, dynamic> json) =>
       _$$_UploadedMediaDataFromJson(json);
@@ -181,15 +197,15 @@ class _$_UploadedMediaData implements _UploadedMediaData {
   @override
   final DateTime expiresAt;
 
-  /// Indicates the language of this media.
+  /// The locale of this media.
   ///
-  /// This field is only set if a caption (.srt) is uploaded.
+  ///
   @override
-  final Language? language;
+  final Locale? locale;
 
   @override
   String toString() {
-    return 'UploadedMediaData(mediaId: $mediaId, id: $id, expiresAt: $expiresAt, language: $language)';
+    return 'UploadedMediaData(mediaId: $mediaId, id: $id, expiresAt: $expiresAt, locale: $locale)';
   }
 
   @override
@@ -200,7 +216,7 @@ class _$_UploadedMediaData implements _UploadedMediaData {
             const DeepCollectionEquality().equals(other.mediaId, mediaId) &&
             const DeepCollectionEquality().equals(other.id, id) &&
             const DeepCollectionEquality().equals(other.expiresAt, expiresAt) &&
-            const DeepCollectionEquality().equals(other.language, language));
+            const DeepCollectionEquality().equals(other.locale, locale));
   }
 
   @JsonKey(ignore: true)
@@ -210,7 +226,7 @@ class _$_UploadedMediaData implements _UploadedMediaData {
       const DeepCollectionEquality().hash(mediaId),
       const DeepCollectionEquality().hash(id),
       const DeepCollectionEquality().hash(expiresAt),
-      const DeepCollectionEquality().hash(language));
+      const DeepCollectionEquality().hash(locale));
 
   @JsonKey(ignore: true)
   @override
@@ -234,7 +250,7 @@ abstract class _UploadedMediaData implements UploadedMediaData {
       @JsonKey(name: 'media_id_string')
           required final String id,
       required final DateTime expiresAt,
-      final Language? language}) = _$_UploadedMediaData;
+      final Locale? locale}) = _$_UploadedMediaData;
 
   factory _UploadedMediaData.fromJson(Map<String, dynamic> json) =
       _$_UploadedMediaData.fromJson;
@@ -256,10 +272,10 @@ abstract class _UploadedMediaData implements UploadedMediaData {
   DateTime get expiresAt;
   @override
 
-  /// Indicates the language of this media.
+  /// The locale of this media.
   ///
-  /// This field is only set if a caption (.srt) is uploaded.
-  Language? get language;
+  ///
+  Locale? get locale;
   @override
   @JsonKey(ignore: true)
   _$$_UploadedMediaDataCopyWith<_$_UploadedMediaData> get copyWith =>